Fix OMP bug in pmat. mapmaking.Signals. Gain-fit
Fixed a nasty bug where the logic for when atomic operations
should be applied in the tod2map operations in the pointing
matrix was inverted. This led to clobbering of pixel values.
In theory this is a nasty bug. In practice it seems to have
had basically no effect.

class Signals and Eqsys2 in mapmaking, which factor out
multi-signal support from the mapmaking equation.
Preconditioners now have a build() function that can be used
to finilize construction once data is available, instead of
doing it all at once.

Work on gain-fitting support in mapmaking, with SignalGain etc.

np.bool -> bool
lmin in dory.fit_src_amps

Support very short TOD lengths in scan.py